[docs] Only server-side render the popular languages
