+++ docs/next.config.js
@@ -4,7 +4,7 @@ const { BundleAnalyzerPlugin } = require('webpack-bundle-analyzer');
 const withTypescript = require('@zeit/next-typescript');
 const pkg = require('../package.json');
 const { findPages } = require('./src/modules/utils/find');
-const { LANGUAGES } = require('./src/modules/constants');
+const { LANGUAGES_SSR } = require('./src/modules/constants');
 
 const workspaceRoot = path.join(__dirname, '../');
 
@@ -140,13 +140,13 @@ module.exports = withTypescript({
     }
 
     // We want to speed-up the build of pull requests.
-    if (process.env.PULL_REQUEST === 'true') {
-      traverse(pages, 'en');
-    } else {
-      LANGUAGES.forEach(userLanguage => {
-        traverse(pages, userLanguage);
-      });
-    }
+    // if (process.env.PULL_REQUEST === 'true') {
+    //   traverse(pages, 'en');
+    // } else {
+    LANGUAGES_SSR.forEach(userLanguage => {
+      traverse(pages, userLanguage);
+    });
+    // }
 
     return map;
   },
