+++ docs/src/modules/components/Ad.js
@@ -52,9 +52,7 @@ function Ad(props) {
   const { classes } = props;
   const { current: random } = React.useRef(Math.random());
   const timerAdblock = React.useRef();
-  const { t } = useSelector(state => ({
-    t: state.options.t,
-  }));
+  const t = useSelector(state => state.options.t);
   const [adblock, setAdblock] = React.useState(null);
 
   const checkAdblock = React.useCallback((attempt = 1) => {
