+++ docs/src/modules/components/DemoLanguages.js
@@ -20,9 +20,7 @@ const styles = {
 function DemoLanguages(props) {
   const { classes, codeOpen, codeVariant, demo, gaEventCategory, onLanguageClick } = props;
   const hasTSVariant = demo.rawTS;
-  const { t } = useSelector(state => ({
-    t: state.options.t,
-  }));
+  const t = useSelector(state => state.options.t);
 
   function renderedCodeVariant() {
     if (codeVariant === CODE_VARIANTS.TS && hasTSVariant) {
