+++ docs/src/modules/components/HomeSteps.js
@@ -105,7 +105,7 @@ const PremiumThemesLink = React.forwardRef((props, ref) => {
 
 function HomeSteps(props) {
   const { classes } = props;
-  const { t } = useSelector(state => ({ t: state.options.t }));
+  const t = useSelector(state => state.options.t);
   const theme = useTheme();
 
   return (
