+++ docs/src/modules/components/HomeUsers.js
@@ -57,7 +57,7 @@ const styles = theme => ({
 
 function HomeUsers(props) {
   const { classes } = props;
-  const { t } = useSelector(state => ({ t: state.options.t }));
+  const t = useSelector(state => state.options.t);
 
   return (
     <div className={classes.root}>
