+++ docs/src/modules/components/Link.js
@@ -36,7 +36,7 @@ function Link(props) {
   } = props;
   const router = useRouter();
 
-  const { userLanguage } = useSelector(state => ({ userLanguage: state.options.userLanguage }));
+  const userLanguage = useSelector(state => state.options.userLanguage);
   const className = clsx(classNameProps, {
     [activeClassName]: router.pathname === props.href && activeClassName,
   });
