+++ docs/src/pages/getting-started/templates/Templates.js
@@ -108,7 +108,7 @@ function layouts(t) {
 
 function Templates(props) {
   const { classes } = props;
-  const { t } = useSelector(state => ({ t: state.options.t }));
+  const t = useSelector(state => state.options.t);
 
   return (
     <Grid container spacing={2}>
